Renamed season.types.ts to season.ts
